Resolves #9. Refactored the actual image generation away from the event listener
